Building mesh change by health
The building in the level will change mesh depending on it's health integer value. The building's health is checked in a function which is placed in a fixed update. For testing, the user can decrease the building's health value by one integer when they press 'A'

If it is at a value of 3 or higher then it is "undamaged"

If it is at a value between 1 and 2 then it will change into a "damaged" mesh, and play a simple placeholder smoke particle, which can be replaced by another particle object.

If it is at a value of 0 or less than 0 then it will change into a "ruined" mesh, and play a simple placeholder explosion and fire particle, which can be replaced by another particle object.